+++ app/models/presenters/facet_values/visibility.rb
@@ -1,7 +1,0 @@
-class Presenters::FacetValues::Visibility < Presenters::FacetValues::URIValue
-
-  def display
-    translate_uri(:visibility, @value)
-  end
-
-end