+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and releases in Jupiter project adheres to [Semantic Versioning](http://semver.o
 ### Added
 - Mounted Oaisys engine
 - Added and set up papertrail gem
+- Added Draper and re-organized facet presenters
 
 ### Changed
 - bump rubocop-rails to 2.4.1 Rails/FilePath default changed to slashes [PR#1398](https://github.com/ualbertalib/jupiter/pull/1398)